Break out core data structures
