+++ src/identifiers.js
@@ -1,0 +1,25 @@
+var numeric = /^[0-9]+$/
+function compareIdentifiers (a, b) {
+  var anum = numeric.test(a)
+  var bnum = numeric.test(b)
+
+  if (anum && bnum) {
+    a = +a
+    b = +b
+  }
+
+  return a === b ? 0
+    : (anum && !bnum) ? -1
+    : (bnum && !anum) ? 1
+    : a < b ? -1
+    : 1
+}
+
+function rcompareIdentifiers (a, b) {
+  return compareIdentifiers(b, a)
+}
+
+module.exports = {
+    compareIdentifiers,
+    rcompareIdentifiers
+};
